Fix Use Better API for Empty String Check
In file: FileDescriptorUtils.java, there is a method fileDescriptorToProtoFile that checks if a {String} is empty using {equals} method. It is better to do the emptiness check using String.isEmpty method as it has less overhead than String.equals method. iCR fixed the issue by using the {String.isEmpty()} method.